Better packfile negotiation, still not great
This verifies that we have all of the commits we say we have,
which is a start, at least. It also actually says we had the last
commit on a branch that changed (if we do), which should help
somewhat with the amount of data downloaded for deltas.

Note that this still doesn't:

* Verify we have everything necessary to reconstruct commits we
  claim to have. That would require tracking that information.
* Follow the git packfile negotiation tactic of sending recent
  branch heads first and checking the responses. This, too, would
  require additional information.